Add files via upload
The data set comprised of 42,535 loan listings spanning from June, 2007 to December, 2011 for customers of the Lending Club, a peer-to-peer lending company that is headquartered in the United States. Established in 2006, investors have the option of browsing loan listings and selecting loans that they want to invest in, based upon the information supplied about the borrower, the loan amount, the loan grade, and the purpose for which the loan has been availed. Investors make money from the interest on the loan. Borrowers are charged an origination fee and investors a service fee.